Extract common parser stuff to parent module
