+++ lib/Intlc/Parser/ICU.hs
@@ -8,14 +8,8 @@
 module Intlc.Parser.ICU where
 
 import qualified Control.Applicative.Combinators.NonEmpty as NE
-import           Data.Aeson                               (decode)
-import           Data.ByteString.Lazy                     (ByteString)
-import qualified Data.Map                                 as M
 import qualified Data.Text                                as T
-import           Data.Validation                          (toEither,
-                                                           validationNel)
 import           Data.Void                                ()
-import           Intlc.Core
 import           Intlc.ICU
 import           Prelude                                  hiding (ByteString)
 import           Text.Megaparsec                          hiding (State, Stream,
@@ -24,13 +18,6 @@ import           Text.Megaparsec                          hiding (State, Stream,
 import           Text.Megaparsec.Char
 import qualified Text.Megaparsec.Char.Lexer               as L
 import           Text.Megaparsec.Error.Builder
-
-type ParseErr = ParseErrorBundle Text MessageParseErr
-
-data ParseFailure
-  = FailedJsonParse
-  | FailedDatasetParse (NonEmpty ParseErr)
-  deriving (Show, Eq)
 
 data MessageParseErr
   = NoClosingCallbackTag Text
@@ -43,20 +30,6 @@ instance ShowErrorComponent MessageParseErr where
 
 failingWith :: MonadParsec e s m => Int -> e -> m a
 pos `failingWith` e = parseError . errFancy pos . fancy . ErrorCustom $ e
-
-printErr :: ParseFailure -> String
-printErr FailedJsonParse         = "Failed to parse JSON"
-printErr (FailedDatasetParse es) = intercalate "\n" . toList . fmap errorBundlePretty $ es
-
-parseDataset :: ByteString -> Either ParseFailure (Dataset Translation)
-parseDataset = parse' <=< decode'
-  where decode' = maybeToRight FailedJsonParse . decode
-        parse' = toEither . first FailedDatasetParse . M.traverseWithKey ((validationNel .) . parseTranslationFor)
-
-parseTranslationFor :: Text -> UnparsedTranslation -> Either ParseErr Translation
-parseTranslationFor name (UnparsedTranslation umsg be md) = do
-  msg' <- runParser (runReaderT msg initialState) (T.unpack name) umsg
-  pure $ Translation msg' be md
 
 data ParserState = ParserState
   { pluralCtxName :: Maybe Text
